Update AccessStudents.py
Hey, looking at the code when i run it says there is no database created for this code. Looking through I cant find where you have the cur.execute("Create TABLE IF NOT EXISTS database") So when the code runs it errors out. This my solution to it, give it a try. Also we need to add a loop that if you add or delete / edit you have the option to add/edit/delete again. Right now the code ends making you only able to make a single edit or change.